refactor(tooling): split check-tsdoc, pre-push, check-exports into internal modules
Split check-tsdoc.ts (750→101), pre-push.ts (567→242),
check-exports.ts (446→168). Internal modules for types,
analysis, formatting, checks, and filesystem helpers.
Add cli/internal/* exclusion to bunup config.

🤘🏻 In-collaboration-with: [Claude Code](https://claude.com/claude-code)
